requirements.txt
djangorestframework==3.13.1
django==4.0.3
psycopg2-binary